feat: Implement all menu item API routes. TODO: fix filter implementation

+++ src/backend/src/main/java/com/example/cs1530/dto/menuitem/CreateMenuItemRequest.java
@@ -1,6 +1,5 @@
 package com.example.cs1530.dto.menuitem;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -13,7 +12,7 @@ public class CreateMenuItemRequest {
     private String description;
 
     @Schema(description = "Price of the menu item", required = true)
-    private BigDecimal price;
+    private Double price;
 
     @Schema(description = "Set of category IDs that the menu item belongs to")
     private Set<Long> categoryIds;
@@ -26,7 +25,7 @@ public class CreateMenuItemRequest {
         return description;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
